13: extract interfaces from `App.tsx` into `types.ts` + extract "constants" from `App.tsx` into `constants.ts`
